more changes

+++ gamestore/src/test/java/com/company/gamestore/controller/ShirtControllerTest.java
@@ -15,6 +15,8 @@ import org.springframework.test.context.junit4.SpringRunner;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
 
+import java.math.BigDecimal;
+
 import static org.junit.Assert.*;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.delete;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.put;
@@ -39,7 +41,7 @@ public class ShirtControllerTest {
         shirt.setColor("Green");
         shirt.setDescription("Netflix Swag");
         shirt.setQuantity(44);
-        shirt.setPrice(20.49);
+        shirt.setPrice(BigDecimal.valueOf(20.49));
         shirt.setSize("Medium");
 
         String inputJson = objectMapper.writeValueAsString(shirt);
@@ -54,7 +56,7 @@ public class ShirtControllerTest {
         shirt.setColor("Green");
         shirt.setDescription("Netflix Swag");
         shirt.setQuantity(44);
-        shirt.setPrice(20.49);
+        shirt.setPrice(BigDecimal.valueOf(20.49));
         shirt.setSize("Medium");
         shirt.setShirtId(1);
         shirtRepo.save(shirt);
@@ -69,7 +71,7 @@ public class ShirtControllerTest {
         shirt.setColor("Green");
         shirt.setDescription("Netflix Swag");
         shirt.setQuantity(44);
-        shirt.setPrice(20.49);
+        shirt.setPrice(BigDecimal.valueOf(20.49));
         shirt.setSize("Medium");
         shirt.setShirtId(1);
         shirtRepo.save(shirt);
@@ -77,7 +79,7 @@ public class ShirtControllerTest {
         shirt.setColor("Red");
         shirt.setDescription("Amazon Swag");
         shirt.setQuantity(43);
-        shirt.setPrice(20.42);
+        shirt.setPrice(BigDecimal.valueOf(20.42));
         shirt.setSize("Small");
 
         shirtRepo.save(shirt);
@@ -98,7 +100,7 @@ public class ShirtControllerTest {
         shirt.setColor("Green");
         shirt.setDescription("Netflix Swag");
         shirt.setQuantity(44);
-        shirt.setPrice(20.49);
+        shirt.setPrice(BigDecimal.valueOf(20.49));
         shirt.setSize("Medium");
         shirtRepo.save(shirt);
         String inputJson = objectMapper.writeValueAsString(shirt);
@@ -111,21 +113,21 @@ public class ShirtControllerTest {
         shirt1.setColor("Green");
         shirt1.setDescription("Netflix Swag");
         shirt1.setQuantity(44);
-        shirt1.setPrice(20.49);
+        shirt1.setPrice(BigDecimal.valueOf(20.49));
         shirt1.setSize("Medium");
 
         Shirt shirt2 = new Shirt();
         shirt2.setColor("Red");
         shirt2.setDescription("Amazon Swag");
         shirt2.setQuantity(43);
-        shirt2.setPrice(20.42);
+        shirt2.setPrice(BigDecimal.valueOf(20.42));
         shirt2.setSize("Small");
 
         Shirt shirt3 = new Shirt();
         shirt3.setColor("Blue");
         shirt3.setDescription("Google Swag");
         shirt3.setQuantity(78);
-        shirt3.setPrice(23.70);
+        shirt3.setPrice(BigDecimal.valueOf(23.70));
         shirt3.setSize("Large");
 
         shirtRepo.save(shirt1);
@@ -143,7 +145,7 @@ public class ShirtControllerTest {
         shirt.setColor("Blue");
         shirt.setDescription("Google Swag");
         shirt.setQuantity(78);
-        shirt.setPrice(23.70);
+        shirt.setPrice(BigDecimal.valueOf(23.70));
         shirt.setSize("Large");
 
         shirtRepo.save(shirt);
@@ -159,7 +161,7 @@ public class ShirtControllerTest {
         shirt.setColor("Blue");
         shirt.setDescription("Google Swag");
         shirt.setQuantity(78);
-        shirt.setPrice(23.70);
+        shirt.setPrice(BigDecimal.valueOf(23.70));
         shirt.setSize("Large");
 
         shirtRepo.save(shirt);
