+++ gamestore/src/test/java/com/company/gamestore/repository/ShirtRepositoryTest.java
@@ -11,6 +11,7 @@ import org.springframework.http.MediaType;
 import org.springframework.test.context.junit4.SpringRunner;
 import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Optional;
 
@@ -38,7 +39,7 @@ public class ShirtRepositoryTest {
         shirt.setColor("Green");
         shirt.setDescription("Netflix Swag");
         shirt.setQuantity(44);
-        shirt.setPrice(20.49);
+        shirt.setPrice(BigDecimal.valueOf(20.49));
         shirt.setSize("Medium");
 
         shirt = shirtRepo.save(shirt);
@@ -52,7 +53,7 @@ public class ShirtRepositoryTest {
         shirt.setColor("Green");
         shirt.setDescription("Netflix Swag");
         shirt.setQuantity(44);
-        shirt.setPrice(20.49);
+        shirt.setPrice(BigDecimal.valueOf(20.49));
         shirt.setSize("Medium");
         shirt = shirtRepo.save(shirt);
 
@@ -69,7 +70,7 @@ public class ShirtRepositoryTest {
         shirt.setColor("Green");
         shirt.setDescription("Netflix Swag");
         shirt.setQuantity(44);
-        shirt.setPrice(20.49);
+        shirt.setPrice(BigDecimal.valueOf(20.49));
         shirt.setSize("Medium");
         shirt.setShirtId(1);
         shirt = shirtRepo.save(shirt);
@@ -77,7 +78,7 @@ public class ShirtRepositoryTest {
         shirt.setColor("Red");
         shirt.setDescription("Amazon Swag");
         shirt.setQuantity(43);
-        shirt.setPrice(20.42);
+        shirt.setPrice(BigDecimal.valueOf(20.42));
         shirt.setSize("Small");
 
         shirt = shirtRepo.save(shirt);
@@ -92,7 +93,7 @@ public class ShirtRepositoryTest {
         shirt.setColor("Green");
         shirt.setDescription("Netflix Swag");
         shirt.setQuantity(44);
-        shirt.setPrice(20.49);
+        shirt.setPrice(BigDecimal.valueOf(20.49));
         shirt.setSize("Medium");
         shirt = shirtRepo.save(shirt);
         Optional<Shirt> shirtOptional = shirtRepo.findById(shirt.getShirtId());
@@ -106,21 +107,21 @@ public class ShirtRepositoryTest {
         shirt1.setColor("Green");
         shirt1.setDescription("Netflix Swag");
         shirt1.setQuantity(44);
-        shirt1.setPrice(20.49);
+        shirt1.setPrice(BigDecimal.valueOf(20.49));
         shirt1.setSize("Medium");
 
         Shirt shirt2 = new Shirt();
         shirt2.setColor("Red");
         shirt2.setDescription("Amazon Swag");
         shirt2.setQuantity(43);
-        shirt2.setPrice(20.42);
+        shirt2.setPrice(BigDecimal.valueOf(20.42));
         shirt2.setSize("Small");
 
         Shirt shirt3 = new Shirt();
         shirt3.setColor("Blue");
         shirt3.setDescription("Google Swag");
         shirt3.setQuantity(78);
-        shirt3.setPrice(23.70);
+        shirt3.setPrice(BigDecimal.valueOf(23.70));
         shirt3.setSize("Large");
 
         shirtRepo.save(shirt1);
@@ -137,7 +138,7 @@ public class ShirtRepositoryTest {
         shirt.setColor("Blue");
         shirt.setDescription("Google Swag");
         shirt.setQuantity(78);
-        shirt.setPrice(23.70);
+        shirt.setPrice(BigDecimal.valueOf(23.70));
         shirt.setSize("Large");
         shirt = shirtRepo.save(shirt);
 
@@ -147,7 +148,7 @@ public class ShirtRepositoryTest {
         shirt2.setColor("Red");
         shirt2.setDescription("Amazon Swag");
         shirt2.setQuantity(43);
-        shirt2.setPrice(20.42);
+        shirt2.setPrice(BigDecimal.valueOf(20.42));
         shirt2.setSize("Small");
 
         shirt2 = shirtRepo.save(shirt2);
@@ -156,7 +157,7 @@ public class ShirtRepositoryTest {
         shirt3.setColor("Blue");
         shirt3.setDescription("Google Swag");
         shirt3.setQuantity(78);
-        shirt3.setPrice(23.70);
+        shirt3.setPrice(BigDecimal.valueOf(23.70));
         shirt3.setSize("Large");
 
         shirt3 = shirtRepo.save(shirt3);
@@ -173,7 +174,7 @@ public class ShirtRepositoryTest {
         shirt.setColor("Blue");
         shirt.setDescription("Google Swag");
         shirt.setQuantity(78);
-        shirt.setPrice(23.70);
+        shirt.setPrice(BigDecimal.valueOf(23.70));
         shirt.setSize("Large");
         shirt = shirtRepo.save(shirt);
 
@@ -183,7 +184,7 @@ public class ShirtRepositoryTest {
         shirt2.setColor("Red");
         shirt2.setDescription("Amazon Swag");
         shirt2.setQuantity(43);
-        shirt2.setPrice(20.42);
+        shirt2.setPrice(BigDecimal.valueOf(20.42));
         shirt2.setSize("Small");
 
         shirt2 = shirtRepo.save(shirt2);
@@ -192,7 +193,7 @@ public class ShirtRepositoryTest {
         shirt3.setColor("Blue");
         shirt3.setDescription("Google Swag");
         shirt3.setQuantity(78);
-        shirt3.setPrice(23.70);
+        shirt3.setPrice(BigDecimal.valueOf(23.70));
         shirt3.setSize("Large");
 
         shirt3 = shirtRepo.save(shirt3);
